WIP: create mock/utils.js
