+++ src/mock/film.js
@@ -1,58 +1,13 @@
 import {nanoid} from 'nanoid';
-import {generateComments} from "./comments";
-import {getRandomInteger} from "./random";
+import {generateComments} from './comments';
+import {getRandomInteger} from './utils';
+import {generateRating} from './utils';
+import {generateInfo} from './utils';
+import {choise} from './utils';
+import {generateDescription} from './utils';
+import {generateTime} from './utils';
+import {generateRandomComments} from './utils';
 
-/**
- * Функция генерация рейтинга (рандомная генерация)
- * @param {number} a - начало промежутка
- * @param {number} b - конец промежутка
- * @return {number}
- */
-const generateRating = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return (lower + Math.random() * (upper - lower + 1)).toFixed(1).replace(`,`, `.`);
-};
-
-/**
- * Функция генерации информации о фильме
- * Определяет объект info и случайным образом выбирает из него один элемент
- * @return {object}
- */
-const generateInfo = () => {
-  const info = [
-    {
-      title: `The Dance of Life`,
-      originTitle: `The Great Flamarion`,
-      poster: `the-dance-of-life.jpg`,
-    },
-    {
-      title: `Sagebrush Trail`,
-      originTitle: `Sagebrush Trail`,
-      poster: `sagebrush-trail.jpg`,
-    },
-    {
-      title: `The Man with the Golden Arm`,
-      originTitle: `The Great Flamarion`,
-      poster: `made-for-each-other.png`,
-    },
-    {
-      title: `Santa Claus Conquers the Martians`,
-      originTitle: `Santa Claus`,
-      poster: `santa-claus-conquers-the-martians.jpg`,
-    },
-    {
-      title: `Popeye the Sailor Meets Sindbad the Sailor`,
-      originTitle: `Popeye the Sailor Meets Sindbad the Sailor`,
-      poster: `popeye-meets-sinbad.png`,
-    }
-  ];
-
-  const randomIndex = getRandomInteger(0, info.length - 1);
-
-  return info[randomIndex];
-};
 
 const regisseurs = [
   `Anthony Mann`,
@@ -93,58 +48,6 @@ const actors = [
   `Vincent Cassel`,
   `Viggo Mortensen`,
 ];
-
-/**
- * Функция создания массивов (выборки из имеющегося случайным образом)
- * @param {Array} arr - входящий массив откуда надо взять данные
- * @param {number} n - количество элементов, которые надо "набрать"
- * @return {Array}
- */
-const choise = (arr, n) => {
-  return new Array(n).fill(null).map(() => arr[getRandomInteger(0, arr.length - 1)]);
-};
-
-/**
- * Функция создания описания фильма (выборки из имеющейся строки подстроки случайным образом)
- * @return {string}
- */
-const generateDescription = () => {
-  const sentences = [`Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
-    `Cras aliquet varius magna, non porta ligula feugiat eget.`,
-    `Fusce tristique felis at fermentum pharetra.`,
-    `Aliquam id orci ut lectus varius viverra.`,
-    `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.`,
-    `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
-    `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`,
-    `Sed sed nisi sed augue convallis suscipit in sed felis.`,
-    `Aliquam erat volutpat.`,
-    `Nunc fermentum tortor ac porta dapibus.`,
-    `In rutrum ac purus sit amet tempus.`];
-
-  return sentences.slice(getRandomInteger(1, sentences.length - 1)).join(` `);
-};
-
-/**
- * Функция генерация времени случайным образом
- * @return {string}
- */
-const generateTime = () => {
-
-  const randomHour = getRandomInteger(0, 3);
-  const randomMinutes = getRandomInteger(0, 59);
-
-  return randomHour + `h ` + randomMinutes + `m`;
-};
-
-/**
- * Функция генерация времени (dayjs) + рандом
- * @return {string}
- */
-const generateRandomComments = () => {
-  const COMMENT_COUNT = getRandomInteger(0, 5);
-  const randomComments = new Array(COMMENT_COUNT).fill(null).map((index) => generateComments(index));
-  return randomComments;
-};
 
 /**
  * Функция создания объекта фильма
